move set() to new file
